+++ config.py
@@ -20,8 +20,8 @@ import os
 
 
 ############################################################
-API_TOKEN = ' '  # TG API BOT TOKEN
-ALLOWED_USER_ID =    # TG USER ID
+API_TOKEN = ' '   # TG API BOT TOKEN
+ALLOWED_USER_ID =                                      # TG USER ID
 ############################################################
